backend/requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.2
attrs==25.3.0
certifi==2025.8.3
Django==5.2.6
django-ckeditor==5.0.0
django-cors-headers==4.9.0
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.17
djangorestframework==3.16.1
drf-yasg==1.21.10
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
magic-filter==1.0.12
multidict==6.6.4
packaging==25.0
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
wheel==0.45.1
yarl==1.20.1
